pyproject.toml
[tool.poetry]
name = "toasty"
version = "0.1.0"
description = "GitHub bot for AI code reviews, project management, and security analysis"
authors = ["OWASP BLT Team"]
readme = "README.md"
packages = [{include = "toasty"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
requests = "^2.32.3"
google-generativeai = "^0.8.5"
jsonschema = "^4.24.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.8.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=toasty --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
